[DROOLS-7523] allow parallel evaluation of LHS part of rules with sequential firing of consequences
