+++ drools-core/src/main/java/org/drools/core/RuleBaseConfiguration.java
@@ -16,46 +16,22 @@
 
 package org.drools.core;
 
+import org.drools.base.rule.consequence.ConflictResolver;
+import org.drools.core.runtime.rule.impl.DefaultConsequenceExceptionHandler;
+import org.drools.util.StringUtils;
+import org.kie.api.KieBaseConfiguration;
+import org.kie.api.conf.*;
+import org.kie.api.runtime.rule.ConsequenceExceptionHandler;
+import org.kie.internal.conf.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.Externalizable;
 import java.io.IOException;
 import java.io.ObjectInput;
 import java.io.ObjectOutput;
 import java.util.HashMap;
 import java.util.Map;
-
-import org.drools.base.rule.consequence.ConflictResolver;
-import org.drools.core.runtime.rule.impl.DefaultConsequenceExceptionHandler;
-import org.drools.util.StringUtils;
-import org.kie.api.KieBaseConfiguration;
-import org.kie.api.conf.BetaRangeIndexOption;
-import org.kie.api.conf.ConfigurationKey;
-import org.kie.api.conf.DeclarativeAgendaOption;
-import org.kie.api.conf.EqualityBehaviorOption;
-import org.kie.api.conf.EventProcessingOption;
-import org.kie.api.conf.KieBaseOption;
-import org.kie.api.conf.MultiValueKieBaseOption;
-import org.kie.api.conf.OptionKey;
-import org.kie.api.conf.RemoveIdentitiesOption;
-import org.kie.api.conf.SequentialOption;
-import org.kie.api.conf.SessionsPoolOption;
-import org.kie.api.conf.SingleValueKieBaseOption;
-import org.kie.api.runtime.rule.ConsequenceExceptionHandler;
-import org.kie.internal.conf.AlphaRangeIndexThresholdOption;
-import org.kie.internal.conf.AlphaThresholdOption;
-import org.kie.internal.conf.CompositeConfiguration;
-import org.kie.internal.conf.CompositeKeyDepthOption;
-import org.kie.internal.conf.ConsequenceExceptionHandlerOption;
-import org.kie.internal.conf.ConstraintJittingThresholdOption;
-import org.kie.internal.conf.IndexLeftBetaMemoryOption;
-import org.kie.internal.conf.IndexPrecedenceOption;
-import org.kie.internal.conf.IndexRightBetaMemoryOption;
-import org.kie.internal.conf.MaxThreadsOption;
-import org.kie.internal.conf.MultithreadEvaluationOption;
-import org.kie.internal.conf.SequentialAgendaOption;
-import org.kie.internal.conf.ShareAlphaNodesOption;
-import org.kie.internal.conf.ShareBetaNodesOption;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * RuleBaseConfiguration
@@ -141,10 +117,10 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
 
     private IndexPrecedenceOption indexPrecedenceOption;
 
-    // if "true", rulebase builder will try to split
+    // if parallelism is enabled, rulebase builder will try to split
     // the rulebase into multiple partitions that can be evaluated
     // in parallel by using multiple internal threads
-    private boolean multithread;
+    private ParallelExecutionOption parallelExecution;
     private int     maxThreads;
 
     private ConflictResolver conflictResolver;
@@ -200,8 +176,8 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
 
         setSequential(Boolean.parseBoolean(getPropertyValue(SequentialOption.PROPERTY_NAME, "false")));
 
-        setMultithreadEvaluation(Boolean.parseBoolean(getPropertyValue(MultithreadEvaluationOption.PROPERTY_NAME,
-                                                                                         "false")));
+        setParallelExecution(ParallelExecutionOption.determineParallelExecution(getPropertyValue(ParallelExecutionOption.PROPERTY_NAME,
+                "sequential")));
 
         setMaxThreads( Integer.parseInt( getPropertyValue( MaxThreadsOption.PROPERTY_NAME,
                                                                              "3" ) ) );
@@ -234,7 +210,7 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
         out.writeObject(consequenceExceptionHandler);
         out.writeObject(ruleBaseUpdateHandler);
         out.writeObject(conflictResolver);
-        out.writeBoolean(multithread);
+        out.writeObject(parallelExecution);
         out.writeInt(maxThreads);
         out.writeObject(eventProcessingMode);
         out.writeBoolean(declarativeAgenda);
@@ -263,7 +239,7 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
         consequenceExceptionHandler = (String) in.readObject();
         ruleBaseUpdateHandler = (String) in.readObject();
         conflictResolver = (ConflictResolver) in.readObject();
-        multithread = in.readBoolean();
+        parallelExecution = (ParallelExecutionOption) in.readObject();
         maxThreads = in.readInt();
         eventProcessingMode = (EventProcessingOption) in.readObject();
         declarativeAgenda = in.readBoolean();
@@ -334,8 +310,8 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
             case MaxThreadsOption.PROPERTY_NAME: {
                 return (T) MaxThreadsOption.get(getMaxThreads());
             }
-            case MultithreadEvaluationOption.PROPERTY_NAME: {
-                return (T) (this.multithread ? MultithreadEvaluationOption.YES : MultithreadEvaluationOption.NO);
+            case ParallelExecutionOption.PROPERTY_NAME: {
+                return (T) parallelExecution;
             }
             case DeclarativeAgendaOption.PROPERTY_NAME: {
                 return (T) (this.isDeclarativeAgenda() ? DeclarativeAgendaOption.ENABLED : DeclarativeAgendaOption.DISABLED);
@@ -419,8 +395,8 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
                 setMaxThreads( ( (MaxThreadsOption) option ).getMaxThreads());
                 break;
             }
-            case MultithreadEvaluationOption.PROPERTY_NAME: {
-                setMultithreadEvaluation( ( (MultithreadEvaluationOption) option ).isMultithreadEvaluation());
+            case ParallelExecutionOption.PROPERTY_NAME: {
+                setParallelExecution( (ParallelExecutionOption) option );
                 break;
             }
             case DeclarativeAgendaOption.PROPERTY_NAME: {
@@ -502,8 +478,8 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
                 setRuleBaseUpdateHandler(StringUtils.isEmpty(value) ? "" : value);
                 break;
             }
-            case MultithreadEvaluationOption.PROPERTY_NAME: {
-                setMultithreadEvaluation(StringUtils.isEmpty(value) ? false : Boolean.valueOf(value));
+            case ParallelExecutionOption.PROPERTY_NAME: {
+                setParallelExecution(ParallelExecutionOption.determineParallelExecution(StringUtils.isEmpty(value) ? "sequential" : value));
                 break;
             }
             case MaxThreadsOption.PROPERTY_NAME: {
@@ -575,8 +551,8 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
             case "drools.ruleBaseUpdateHandler": {
                 return getRuleBaseUpdateHandler();
             }
-            case MultithreadEvaluationOption.PROPERTY_NAME: {
-                return Boolean.toString(isMultithreadEvaluation());
+            case ParallelExecutionOption.PROPERTY_NAME: {
+                return parallelExecution.toExternalForm();
             }
             case MaxThreadsOption.PROPERTY_NAME: {
                 return Integer.toString(getMaxThreads());
@@ -762,21 +738,13 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
         this.sequentialAgenda = sequentialAgenda;
     }
 
-    /**
-     * Defines if the RuleBase should be executed using a pool of
-     * threads for evaluating the rules ("true"), or if the rulebase 
-     * should work in classic single thread mode ("false").
-     * 
-     * @param enableMultithread true for multi-thread or 
-     *                     false for single-thread. Default is false.
-     */
-    public void setMultithreadEvaluation(boolean enableMultithread) {
+    public void setParallelExecution(ParallelExecutionOption parallelExecutionOption) {
         checkCanChange();
-        this.multithread = enableMultithread;
+        this.parallelExecution = parallelExecutionOption;
     }
 
     public void enforceSingleThreadEvaluation() {
-        this.multithread = false;
+        this.parallelExecution = ParallelExecutionOption.SEQUENTIAL;
     }
 
     /**
@@ -785,8 +753,12 @@ public class RuleBaseConfiguration  extends BaseConfiguration<KieBaseOption, Sin
      * 
      * @return
      */
-    public boolean isMultithreadEvaluation() {
-        return this.multithread;
+    public boolean isParallelEvaluation() {
+        return this.parallelExecution.isParallel();
+    }
+
+    public boolean isParallelExecution() {
+        return this.parallelExecution == ParallelExecutionOption.FULLY_PARALLEL;
     }
 
     /**
