+++ drools-core/src/main/java/org/drools/core/common/InternalAgendaGroup.java
@@ -16,9 +16,10 @@
 
 package org.drools.core.common;
 
+import org.drools.core.phreak.RuleAgendaItem;
+
 import java.util.Collection;
 import java.util.Map;
-import org.drools.core.phreak.RuleAgendaItem;
 
 public interface InternalAgendaGroup extends org.kie.api.runtime.rule.AgendaGroup {
 
@@ -67,8 +68,6 @@ public interface InternalAgendaGroup extends org.kie.api.runtime.rule.AgendaGrou
     void remove(RuleAgendaItem activation);
 
     void setActive(boolean activate);
-
-    Collection<RuleAgendaItem> getAll();
 
     void setActivatedForRecency(long recency);
