+++ drools-kiesession/src/main/java/org/drools/kiesession/agenda/PartitionedDefaultAgenda.java
@@ -17,39 +17,33 @@
 package org.drools.kiesession.agenda;
 
 import org.drools.core.common.InternalFactHandle;
-import org.drools.core.common.InternalWorkingMemoryEntryPoint;
-import org.drools.core.impl.InternalRuleBase;
-import org.drools.core.reteoo.ObjectTypeNode;
+import org.drools.core.common.InternalWorkingMemory;
 import org.drools.core.common.PropagationContext;
+import org.drools.core.reteoo.ObjectTypeNode;
 
 public class PartitionedDefaultAgenda extends DefaultAgenda {
 
     private final int partition;
 
-    PartitionedDefaultAgenda(InternalRuleBase kBase,
-                             boolean initMain,
+    PartitionedDefaultAgenda(InternalWorkingMemory workingMemory,
                              ExecutionStateMachine executionStateMachine,
                              int partition) {
-        super(kBase, initMain, executionStateMachine);
+        super(workingMemory, executionStateMachine);
         this.partition = partition;
     }
 
     /**
      * Do not use this constructor! It should be used just by deserialization.
      */
-    public PartitionedDefaultAgenda() {
-        partition = 0;
+    @Override
+    protected void retractFactHandle(PropagationContext ectx, InternalFactHandle factHandle) {
+        ObjectTypeNode.retractLeftTuples(factHandle, ectx, workingMemory, partition );
+        ObjectTypeNode.retractRightTuples(factHandle, ectx, workingMemory, partition );
     }
 
     @Override
-    protected void doRetract( PropagationContext ectx ) {
-        InternalFactHandle factHandle = (InternalFactHandle) ectx.getFactHandle();
-        ObjectTypeNode.retractLeftTuples( factHandle, ectx, workingMemory, partition );
-        ObjectTypeNode.retractRightTuples( factHandle, ectx, workingMemory, partition );
-        if ( isMainPartition() && factHandle.isPendingRemoveFromStore() ) {
-            String epId = factHandle.getEntryPointName();
-            ( (InternalWorkingMemoryEntryPoint) workingMemory.getEntryPoint( epId ) ).removeFromObjectStore( factHandle );
-        }
+    protected boolean isPendingRemoveFactHandleFromStore(InternalFactHandle factHandle) {
+        return isMainPartition() && super.isPendingRemoveFactHandleFromStore(factHandle);
     }
 
     private boolean isMainPartition() {
