+++ drools-reliability/drools-reliability-core/src/main/java/org/drools/reliability/core/ReliableAgenda.java
@@ -15,8 +15,8 @@
 
 package org.drools.reliability.core;
 
+import org.drools.core.common.InternalWorkingMemory;
 import org.drools.core.common.Storage;
-import org.drools.core.impl.InternalRuleBase;
 import org.drools.core.phreak.PropagationList;
 import org.drools.kiesession.agenda.DefaultAgenda;
 
@@ -24,14 +24,8 @@ import static org.drools.reliability.core.ReliablePropagationList.PROPAGATION_LI
 
 public class ReliableAgenda extends DefaultAgenda {
 
-    public ReliableAgenda() { }
-
-    public ReliableAgenda(InternalRuleBase kBase) {
-        super( kBase );
-    }
-
-    public ReliableAgenda(InternalRuleBase kBase, boolean initMain) {
-        super( kBase, initMain );
+    public ReliableAgenda(InternalWorkingMemory workingMemory) {
+        super( workingMemory );
     }
 
     @Override
