+++ drools-reliability/drools-reliability-core/src/main/java/org/drools/reliability/core/ReliableAgendaFactory.java
@@ -19,7 +19,7 @@ package org.drools.reliability.core;
 
 import org.drools.core.common.AgendaFactory;
 import org.drools.core.common.InternalAgenda;
-import org.drools.core.impl.InternalRuleBase;
+import org.drools.core.common.InternalWorkingMemory;
 
 import java.io.Serializable;
 
@@ -33,7 +33,7 @@ public class ReliableAgendaFactory implements AgendaFactory, Serializable {
 
     private ReliableAgendaFactory() { }
 
-    public InternalAgenda createAgenda(InternalRuleBase kBase, boolean initMain) {
-        return new ReliableAgenda( kBase, initMain );
+    public InternalAgenda createAgenda(InternalWorkingMemory workingMemory) {
+        return new ReliableAgenda( workingMemory );
     }
 }
