+++ drools-serialization-protobuf/src/main/java/org/drools/serialization/protobuf/ProtobufOutputMarshaller.java
@@ -238,7 +238,7 @@ public class ProtobufOutputMarshaller {
             // this must clone as re-evaluation will under underlying Collection
             for ( RuleAgendaItem activation : new ArrayList<>(wm.getAgenda().getAgendaGroupsManager().getActivations())) {
                 // evaluate it
-                activation.getRuleExecutor().reEvaluateNetwork( wm );
+                activation.getRuleExecutor().evaluateNetworkIfDirty( wm );
                 activation.getRuleExecutor().removeRuleAgendaItemWhenEmpty( wm );
             }
             dirty = false;
